refactor(studio-books): split book create and detail pages
