+++ packages/studio/src/pages/BookDetail.tsx
@@ -7,51 +7,27 @@ import { useColors } from "../hooks/use-colors";
 import { deriveBookActivity, shouldRefetchBookView } from "../hooks/use-book-activity";
 import { ConfirmDialog } from "../components/ConfirmDialog";
 import {
-  ChevronLeft,
-  Zap,
   FileText,
   CheckCheck,
   BarChart2,
   Download,
-  Search,
-  Wand2,
-  Eye,
   Database,
   Check,
   X,
   ShieldCheck,
   RotateCcw,
   RefreshCw,
-  Sparkles,
-  Trash2,
   Save
 } from "lucide-react";
-
-interface ChapterMeta {
-  readonly number: number;
-  readonly title: string;
-  readonly status: string;
-  readonly wordCount: number;
-}
-
-interface BookData {
-  readonly book: {
-    readonly id: string;
-    readonly title: string;
-    readonly genre: string;
-    readonly status: string;
-    readonly chapterWordCount: number;
-    readonly targetChapters?: number;
-    readonly language?: string;
-    readonly fanficMode?: string;
-  };
-  readonly chapters: ReadonlyArray<ChapterMeta>;
-  readonly nextChapter: number;
-}
-
-type ReviseMode = "spot-fix" | "polish" | "rewrite" | "rework" | "anti-detect";
-type ExportFormat = "txt" | "md" | "epub";
-type BookStatus = "active" | "paused" | "outlining" | "completed" | "dropped";
+import { BookDetailHeader } from "./book-detail/BookDetailHeader";
+import {
+  STATUS_CONFIG,
+  type BookData,
+  type BookStatus,
+  type ExportFormat,
+  type ReviseMode,
+  translateChapterStatus,
+} from "./book-detail-state";
 
 interface Nav {
   toDashboard: () => void;
@@ -59,26 +35,6 @@ interface Nav {
   toAnalytics: (bookId: string) => void;
   toTruth: (bookId: string) => void;
 }
-
-function translateChapterStatus(status: string, t: TFunction): string {
-  const map: Record<string, () => string> = {
-    "ready-for-review": () => t("chapter.readyForReview"),
-    "approved": () => t("chapter.approved"),
-    "drafted": () => t("chapter.drafted"),
-    "needs-revision": () => t("chapter.needsRevision"),
-    "imported": () => t("chapter.imported"),
-    "audit-failed": () => t("chapter.auditFailed"),
-  };
-  return map[status]?.() ?? status;
-}
-
-const STATUS_CONFIG: Record<string, { color: string; icon: React.ReactNode }> = {
-  "ready-for-review": { color: "text-amber-500 bg-amber-500/10", icon: <Eye size={12} /> },
-  approved: { color: "text-emerald-500 bg-emerald-500/10", icon: <Check size={12} /> },
-  drafted: { color: "text-muted-foreground bg-muted/20", icon: <FileText size={12} /> },
-  "needs-revision": { color: "text-destructive bg-destructive/10", icon: <RotateCcw size={12} /> },
-  imported: { color: "text-blue-500 bg-blue-500/10", icon: <Download size={12} /> },
-};
 
 export function BookDetail({
   bookId,
@@ -303,94 +259,20 @@ export function BookDetail({
 
   return (
     <div className="space-y-8 fade-in">
-      {/* Breadcrumbs */}
-      <nav className="flex items-center gap-2 text-[13px] font-medium text-muted-foreground">
-        <button
-          onClick={nav.toDashboard}
-          className="hover:text-primary transition-colors flex items-center gap-1"
-        >
-          <ChevronLeft size={14} />
-          {t("bread.books")}
-        </button>
-        <span className="text-border">/</span>
-        <span className="text-foreground">{book.title}</span>
-      </nav>
-
-      {/* Header Section */}
-      <div className="flex flex-col md:flex-row md:items-end justify-between gap-6 border-b border-border/40 pb-8">
-        <div className="space-y-2">
-          <div className="flex items-center gap-3">
-            <h1 className="text-4xl font-serif font-medium">{book.title}</h1>
-            {book.language === "en" && (
-              <span className="px-1.5 py-0.5 rounded border border-primary/20 text-primary text-[10px] font-bold">EN</span>
-            )}
-          </div>
-          <div className="flex flex-wrap items-center gap-x-4 gap-y-2 text-sm text-muted-foreground font-medium">
-            <span className="px-2 py-0.5 rounded bg-secondary/50 text-foreground/70 uppercase tracking-wider text-xs">{book.genre}</span>
-            <div className="flex items-center gap-1.5">
-              <FileText size={14} />
-              <span>{chapters.length} {t("dash.chapters")}</span>
-            </div>
-            <div className="flex items-center gap-1.5">
-              <Zap size={14} />
-              <span>{totalWords.toLocaleString()} {t("book.words")}</span>
-            </div>
-            {book.fanficMode && (
-              <span className="flex items-center gap-1 text-purple-500">
-                <Sparkles size={12} />
-                <span className="italic">fanfic:{book.fanficMode}</span>
-              </span>
-            )}
-          </div>
-        </div>
-
-        <div className="flex flex-wrap gap-2">
-          <button
-            onClick={handleWriteNext}
-            disabled={writing || drafting}
-            className="flex items-center gap-2 px-5 py-2.5 text-sm font-bold bg-primary text-primary-foreground rounded-xl hover:scale-105 active:scale-95 transition-all shadow-lg shadow-primary/20 disabled:opacity-50"
-          >
-            {writing ? <div className="w-4 h-4 border-2 border-primary-foreground/20 border-t-primary-foreground rounded-full animate-spin" /> : <Zap size={16} />}
-            {writing ? t("dash.writing") : t("book.writeNext")}
-          </button>
-          <button
-            onClick={handleDraft}
-            disabled={writing || drafting}
-            className="flex items-center gap-2 px-5 py-2.5 text-sm font-bold bg-secondary text-foreground rounded-xl hover:bg-secondary/80 transition-all border border-border/50 disabled:opacity-50"
-          >
-            {drafting ? <div className="w-4 h-4 border-2 border-muted-foreground/20 border-t-muted-foreground rounded-full animate-spin" /> : <Wand2 size={16} />}
-            {drafting ? t("book.drafting") : t("book.draftOnly")}
-          </button>
-          <button
-            onClick={() => setConfirmDeleteOpen(true)}
-            disabled={deleting}
-            className="flex items-center gap-2 px-5 py-2.5 text-sm font-bold bg-destructive/10 text-destructive rounded-xl hover:bg-destructive hover:text-white transition-all border border-destructive/20 disabled:opacity-50"
-          >
-            {deleting ? <div className="w-4 h-4 border-2 border-destructive/20 border-t-destructive rounded-full animate-spin" /> : <Trash2 size={16} />}
-            {deleting ? t("common.loading") : t("book.deleteBook")}
-          </button>
-        </div>
-      </div>
-
-      {(writing || drafting || activity.lastError) && (
-        <div
-          className={`rounded-2xl border px-4 py-3 text-sm ${
-            activity.lastError
-              ? "border-destructive/30 bg-destructive/5 text-destructive"
-              : "border-primary/20 bg-primary/[0.04] text-foreground"
-          }`}
-        >
-          {activity.lastError ? (
-            <span>
-              {t("book.pipelineFailed")}: {activity.lastError}
-            </span>
-          ) : writing ? (
-            <span>{t("book.pipelineWriting")}</span>
-          ) : (
-            <span>{t("book.pipelineDrafting")}</span>
-          )}
-        </div>
-      )}
+      <BookDetailHeader
+        book={book}
+        chapterCount={chapters.length}
+        totalWords={totalWords}
+        writing={writing}
+        drafting={drafting}
+        deleting={deleting}
+        lastError={activity.lastError ?? undefined}
+        onBack={nav.toDashboard}
+        onWriteNext={handleWriteNext}
+        onDraft={handleDraft}
+        onDeleteClick={() => setConfirmDeleteOpen(true)}
+        t={t}
+      />
 
       {/* Tool Strip */}
       <div className="flex flex-wrap items-center gap-2 py-1">
